wha_api/requirements.txt
python-socketio
aiohttp
asyncio
python-engineio 